# Improve updateReadGroupComms performance