--- a/src/HMI/lux-data-bind.js
+++ b/src/HMI/lux-data-bind.js
@@ -1083,9 +1083,13 @@ LUX.addVarWriteEvents = function () {
 
 };
 LUX.updateReadGroupComms = function () {
+	if(readGroups.length == 1 && readGroups[0] === 'global') {
+		// This is a short circuit to save time, getting read group elements is expensive
+		return;
+	}
 	const readGroupElements = new Set(Array.from(document.querySelectorAll('[data-read-group]')).map(el => el.getAttribute('data-read-group')));
 	LUX.machines.forEach(machine => {
-		machine.getReadGroupList().forEach(ReadGroupName => {
+		readGroups.forEach(ReadGroupName => {
 			if (ReadGroupName !== 'global') {
 				machine.readGroupShouldManage(ReadGroupName, readGroupElements.has(ReadGroupName));
 			}