Tested with 27 read groups I found that the performance was: Groups: 27 numTests: 1000
new: 0.3977750001940876ms
old: 7.823864999925718ms

Groups: 20 numTests: 1000
new: 0.371430000057444ms
old: 5.58591500017792ms

Groups: 15 numTests: 1000
new: 0.4374599999282509ms
old: 3.727479999884963ms

Groups: 10 numTests: 1000
new: 0.4364149998873472ms
old: 2.6039450000971556ms

Groups: 5 numTests: 1000
new: 0.3595650000497699ms
old: 1.1045599998906255ms

Groups: 2 numTests: 1000
new: 0.3924499999266118ms
old: 0.3316500000655651ms

Groups: 1 numTests: 1000 (global)
new: 0.37988499994389713ms
old: 0.00045000016689300537ms

## What:
Describe what the changes are.

## Why:
Describe the improvement(s) that these changes bring. If applicable, describe the reasoning behind key implementation details.